switching number encoding for faster performance
the previous version used a weird version based on toCharCode / charCodeAt, which is somewhat slow. by switching to parseInt/toString, we can gain approx. 10% performance

note that it only works up until zoom level 10 before it breaks, results in an inflated file size (by approx. 20%), but since that is our default setting, this is no problem